Linter errors fixed
